updated peertable to use hash - implemented updatePeer

+++ src/peertable/PeerRecord.java
@@ -12,17 +12,20 @@ public class PeerRecord {
 	String peerID;
 	InetAddress peerIPAddress;
 	int peerSeqNum;
-	int expirationTime;
+	/**
+	 * peer expiration time, in ms
+	 */
+	long expirationTime;
 	PeerState peerState;
 
-	public PeerRecord(String peerID, InetAddress peerIPAddress, int peerSeqNum, int expirationTime, PeerState peerState) {
+	public PeerRecord(String peerID, InetAddress peerIPAddress, int peerSeqNum, long expirationTime, PeerState peerState) {
 		this.peerID = peerID;
 		this.peerIPAddress = peerIPAddress;
 		this.peerSeqNum = peerSeqNum;
 		this.expirationTime = expirationTime;
 		this.peerState = peerState;
 	}
-
+	
 	@Override
 	public String toString() {
 		return "PeerRecord{" + "peerID=" + peerID + ", peerIPAddress=" + peerIPAddress + ", peerSeqNum=" + peerSeqNum + ", expirationTime=" + expirationTime + ", peerState=" + peerState + '}';
